requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile .requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via httpx
blis==1.3.0
    # via thinc
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via typer
cloudpathlib==0.21.1
    # via weasel
confection==0.1.5
    # via
    #   thinc
    #   weasel
cymem==2.0.11
    # via
    #   preshed
    #   spacy
    #   thinc
dateparser==1.2.2
    # via -r .requirements.in
ftfy==6.3.1
    # via -r .requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via ollama
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via spacy
langcodes==3.5.0
    # via spacy
language-data==1.3.0
    # via langcodes
marisa-trie==1.3.1
    # via language-data
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
murmurhash==1.0.13
    # via
    #   preshed
    #   spacy
    #   thinc
natsort==8.4.0
    # via -r .requirements.in
numpy==2.3.2
    # via
    #   blis
    #   spacy
    #   thinc
ollama==0.5.3
    # via -r .requirements.in
packaging==25.0
    # via
    #   spacy
    #   thinc
    #   weasel
preshed==3.0.10
    # via
    #   spacy
    #   thinc
pydantic==2.11.7
    # via
    #   confection
    #   ollama
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
pypdf2==3.0.1
    # via -r .requirements.in
pyspellchecker==0.8.3
    # via -r .requirements.in
python-dateutil==2.9.0.post0
    # via dateparser
pytz==2025.2
    # via dateparser
regex==2025.7.34
    # via dateparser
requests==2.32.5
    # via
    #   spacy
    #   weasel
rich==14.1.0
    # via typer
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
smart-open==7.3.0.post1
    # via weasel
sniffio==1.3.1
    # via anyio
spacy==3.8.7
    # via -r .requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.5.1
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
thinc==8.3.6
    # via spacy
tqdm==4.67.1
    # via spacy
typer==0.16.1
    # via
    #   spacy
    #   weasel
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzlocal==5.3.1
    # via dateparser
urllib3==2.5.0
    # via requests
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
wcwidth==0.2.13
    # via ftfy
weasel==0.4.1
    # via spacy
wrapt==1.17.3
    # via smart-open

# The following packages are considered to be unsafe in a requirements file:
# setuptools
